refactor: split supervisor manager into focused files
Extract registry, control, runner, lifecycle, and config helpers from
manager.go so Manager stays a thin coordinator.
